requirements.txt
flask
Flask-SocketIO
Flask-Cors
pydub
geopy
google-generativeai
SpeechRecognition
